fix: unite validator function into single file and move composeValidators to distinct file
