--- src/helpers/required.tsx
@@ -1,1 +1,0 @@
-export const required = (value: string) => (value ? undefined : "Required");
