+++ src/helpers/validators.tsx
@@ -1,2 +1,4 @@
 export const empty = (value: string) =>
   value.trim() ? undefined : "No empty spaces";
+
+export const required = (value: string) => (value ? undefined : "Required");
