+++ src/utils/composeValidators.tsx
@@ -1,0 +1,7 @@
+export const composeValidators =
+  (...validators: any) =>
+  (value: any) =>
+    validators.reduce(
+      (error: any, validator: any) => error || validator(value),
+      undefined
+    );
